Refactor banner editor to use generic variants editor
